chore(rust): bump str0m (#9852)
The latest version of str0m includes a fix that would result in an
immediate ICE timeout if a remote candidate was added prior to a local
candidate. We mitigated this in #9793 to make Firezone overall more
resilient towards sudden changes in the ICE connection state.

As a defense-in-depth measure, we also fixed this issue in str0m by not
transitioning to `Disconnected` if haven't even formed an candidate
pairs yet.

Diff:
https://github.com/algesten/str0m/compare/2153bf0385c12d09907b872158e15012755fe5f2...3d6e3d2f2745c9e8c561603b99c034c9bab7670f